Fix from epiphyte comments.
