+++ synapse/common.py
@@ -354,7 +354,7 @@ def yamlsave(obj, *paths):
 
 def yamlmod(obj, *paths):
     '''
-    Combines/creates a yaml file and combines with obj.  obj and file must be maps or empty.
+    Combines/creates a yaml file and combines with obj.  obj and file must be maps/dict or empty.
     '''
     oldobj = yamlload(*paths)
     if obj:
